+++ src/mocks/Chat.tsx
@@ -781,6 +781,7 @@ const conversation = {
         name: 'Effie Cormier',
         phone: '9044222334',
         maskedPhone: '9044222334',
+        fields: '{}',
         lastMessageAt: '2020-06-29T09:31:47Z',
         status: 'VALID',
         bspStatus: 'SESSION_AND_HSM',
@@ -832,6 +833,7 @@ const conversationWithMultipleMessages = {
         name: 'Effie Cormier',
         phone: '9044222334',
         maskedPhone: '9044222334',
+        fields: '{}',
         lastMessageAt: '2020-06-29T09:31:47Z',
         status: 'VALID',
         bspStatus: 'SESSION_AND_HSM',
@@ -917,6 +919,7 @@ const searchQueryResult = {
           name: 'Vaibhav',
           lastMessageAt: '2020-06-29T09:31:47Z',
           status: 'VALID',
+          fields: '{}',
           bspStatus: 'SESSION_AND_HSM',
           isOrgRead: true,
         },
